renamed remaining packages in test not following the convention
