--- src/test/java/com/dlb/chess/san/validate/statically/format/package-info.java
@@ -1,2 +1,0 @@
-@org.eclipse.jdt.annotation.NonNullByDefault
-package com.dlb.chess.san.validate.statically.format;
