+++ src/test/java/com/dlb/chess/test/san/format/oracle/TestSanValidateFormatAgainstReferenceForOracle.java
@@ -7,8 +7,8 @@ import org.junit.jupiter.api.Test;
 import com.dlb.chess.common.NonNullWrapperCommon;
 import com.dlb.chess.san.model.SanParse;
 import com.dlb.chess.san.validate.format.SanValidateFormat;
-import com.dlb.chess.san.validate.statically.format.calculate.SanValidateStaticallyFormat;
 import com.dlb.chess.test.san.reference.SanValidateFormatReference;
+import com.dlb.chess.test.san.validate.statically.format.calculate.SanValidateStaticallyFormat;
 
 class TestSanValidateFormatAgainstReferenceForOracle {
