+++ src/test/java/com/dlb/chess/test/san/validate/statically/format/calculate/AbstractSanValidateStaticallyFormatCalculate.java
@@ -1,4 +1,4 @@
-package com.dlb.chess.san.validate.statically.format.calculate;
+package com.dlb.chess.test.san.validate.statically.format.calculate;
 
 import java.util.ArrayList;
 import java.util.List;
