+++ src/test/java/com/dlb/chess/test/san/validate/statically/format/calculate/RookSanValidateStaticallyFormatCalculate.java
@@ -1,4 +1,4 @@
-package com.dlb.chess.san.validate.statically.format.calculate;
+package com.dlb.chess.test.san.validate.statically.format.calculate;
 
 import java.util.Map;
 import java.util.TreeMap;
@@ -8,14 +8,14 @@ import com.dlb.chess.san.model.SanParse;
 import com.dlb.chess.san.model.SanValidationFromTo;
 import com.google.common.collect.ImmutableMap;
 
-public class QueenSanValidateStaticallyFormatCalculate extends AbstractSanValidateStaticallyFormatCalculate {
+public class RookSanValidateStaticallyFormatCalculate extends AbstractSanValidateStaticallyFormatCalculate {
 
   static ImmutableMap<String, SanParse> calculateSanMap() {
 
     final Map<String, SanParse> sanValidateMap = new TreeMap<>();
 
     for (final SanValidationFromTo model : calculateForPiece()) {
-      populateMap(sanValidateMap, model, QUEEN);
+      populateMap(sanValidateMap, model, ROOK);
     }
 
     return NonNullWrapperCommon.copyOfMap(sanValidateMap);
