+++ src/test/java/com/dlb/chess/test/san/validate/statically/format/package-info.java
@@ -1,0 +1,2 @@
+@org.eclipse.jdt.annotation.NonNullByDefault
+package com.dlb.chess.test.san.validate.statically.format;
